refactor(process): migrate submitter component to v2
