+++ comp/process/submitter/fx/fx.go
@@ -1,0 +1,21 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+// Package fx provides the fx module for the submitter component.
+package fx
+
+import (
+	"go.uber.org/fx"
+
+	submitterimpl "github.com/DataDog/datadog-agent/comp/process/submitter/impl"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+)
+
+// Module defines the fx options for this component.
+func Module() fxutil.Module {
+	return fxutil.Component(
+		fx.Provide(submitterimpl.NewComponent),
+	)
+}
