+++ comp/process/submitter/impl/submitter.go
@@ -20,17 +20,10 @@ import (
 	"github.com/DataDog/datadog-agent/comp/process/agent"
 	forwarders "github.com/DataDog/datadog-agent/comp/process/forwarders/def"
 	"github.com/DataDog/datadog-agent/comp/process/hostinfo/def"
-	submitterComp "github.com/DataDog/datadog-agent/comp/process/submitter"
+	submitterComp "github.com/DataDog/datadog-agent/comp/process/submitter/def"
 	"github.com/DataDog/datadog-agent/comp/process/types"
 	processRunner "github.com/DataDog/datadog-agent/pkg/process/runner"
-	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
-
-// Module defines the fx options for this component.
-func Module() fxutil.Module {
-	return fxutil.Component(
-		fx.Provide(newSubmitter))
-}
 
 // submitter implements the Component.
 type submitterImpl struct {
@@ -57,7 +50,8 @@ type result struct {
 	Submitter          submitterComp.Component
 }
 
-func newSubmitter(deps dependencies) (result, error) {
+// NewComponent creates a new submitter component.
+func NewComponent(deps dependencies) (result, error) {
 	s, err := processRunner.NewSubmitter(deps.Config, deps.Log, deps.Forwarders, deps.Statsd, deps.HostInfo.Object().HostName, deps.SysProbeConfig)
 	if err != nil {
 		return result{}, err
