+++ comp/process/submitter/impl/submitter_test.go
@@ -18,7 +18,8 @@ import (
 	connectionsforwardermock "github.com/DataDog/datadog-agent/comp/forwarder/connectionsforwarder/mock"
 	forwardersimpl "github.com/DataDog/datadog-agent/comp/process/forwarders/mock"
 	hostinfomock "github.com/DataDog/datadog-agent/comp/process/hostinfo/mock"
-	"github.com/DataDog/datadog-agent/comp/process/submitter"
+	submitter "github.com/DataDog/datadog-agent/comp/process/submitter/def"
+	submitterfx "github.com/DataDog/datadog-agent/comp/process/submitter/fx"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
 
@@ -31,6 +32,6 @@ func TestSubmitterLifecycle(t *testing.T) {
 		fx.Provide(func() statsd.ClientInterface {
 			return &statsd.NoOpClient{}
 		}),
-		Module(),
+		submitterfx.Module(),
 	))
 }
